backend/requirements.txt
Flask==3.1.2
flask_cors==6.0.1
mysql_connector_repackaged==0.3.1
